Separated PathMode out to a dedicated file
